move append_measurement tests
